# perf(grpc): use vtproto functions as a default codec