### Description

This PR creates a wrapper for a default grpc codec that will use VTproto functions when possible to un/marshal objects. This will reduce usage of reflection. Since we use bytes pool we cannot use unsafe unmarshaller as we do not own the byte slice [see doc](https://github.com/grpc/grpc-go/blob/v1.70.0/mem/buffers.go#L46-L48) (like we do in pg stores).

Reference implementation: https://github.com/grpc/grpc-go/blob/v1.70.0/encoding/proto/proto.go#L33

```python
                       │   old.txt    │               new.txt                │
                       │    sec/op    │    sec/op     vs base                │
ProtoUnmarshal/small-8   46.08µ ±  1%   42.39µ ± 18%        ~ (p=0.143 n=10)
ProtoUnmarshal/big-8     3.489m ± 63%   2.771m ± 64%        ~ (p=0.353 n=10)
geomean                  401.0µ         342.8µ        -14.53%

                       │    old.txt    │               new.txt                │
                       │     B/op      │     B/op       vs base               │
ProtoUnmarshal/small-8   9.401Ki ±  0%   9.401Ki ±  0%       ~ (p=1.000 n=10)
ProtoUnmarshal/big-8     1.946Mi ± 72%   1.944Mi ± 72%       ~ (p=0.853 n=10)
geomean                  136.9Ki         136.8Ki        -0.05%

                       │   old.txt    │                new.txt                │
                       │  allocs/op   │  allocs/op    vs base                 │
ProtoUnmarshal/small-8    168.0 ±  0%    168.0 ±  0%       ~ (p=1.000 n=10) ¹
ProtoUnmarshal/big-8     17.12k ± 63%   17.12k ± 63%       ~ (p=0.927 n=10)
geomean                  1.696k         1.696k        +0.00%

```


- [x] CHANGELOG update is not needed
- [x] Documentation is not needed
### Testing
- [x] inspected CI results
#### Automated testing
- [x] modified existing tests
- [x] contributed **no automated tests**
#### How I validated my change
CI